calculates monthly and projected annual savings

+++ python_introduction/finance_calculator.py
@@ -2,10 +2,10 @@
 
 
 monthly_income = input("Enter your monthly income: ")
-monthly_income = int(monthly_income)
+monthly_income = float(monthly_income)
 
 total_monthly_expense = input("Enter your total monthly expenses :")
-total_monthly_expense = int(total_monthly_expense)
+total_monthly_expense = float(total_monthly_expense)
 
 monthly_savings = monthly_income - total_monthly_expense
 Projected_Savings = monthly_savings * 12 + (monthly_savings * 12 * 0.05)
